refactor: typings update to d.ts
