+++ src/pages/user/detail.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { useUser } from 'src/hooks/use-user';
+import { UserDetail } from 'src/typings/common';
 import Profile from '../../components/user/profile';
 
 export default function UserDetailPage() {
