+++ src/pages/user/index.tsx
@@ -3,6 +3,7 @@ import { useUser } from 'src/hooks/use-user';
 import { useCallback, useEffect, useRef, useState } from 'react';
 import { Modal } from 'src/components/common';
 import { useAlert } from 'src/components/user/use-common';
+import { IUserRes } from 'src/typings/common';
 
 export default function UserPage() {
   const [data, setData] = useState<IUserRes>({
